modified sorting and update method for planning page, added photos
